Create README.md
We used Java language to create a complete
educational application. This project has helped students
and teachers to gain a better usage of technology in
comparison to the traditional teaching style. the teachers
can evaluate the student’s level by letting the students use
this application without the need for paper. also helps in
setting a time for students to meet their teachers by having
accounts in the system. Teachers can access their
accounts and see how many students are registered to see
them. In conclusion, it has devolved our knowledge of
java and its usage in applications.